Use fake server in tests
